Do not skip installation of component if code was changed
Without this patch, the installation of a component could be
skipped if the code has changed, but the user overrides stay the
same.

This is a problem, as code should always trigger a
redeployment.

This fixes it by removing all the conditionals, therefore ensuring
the deployments always happen (therefore removing all races
conditions of failures).

The conditional used to override does not need to exist anymore
and has been removed.