Make the role usable outside automation
If the user doesn't use automation, the kubectl tasks would fail
by trying to run automation to get kubernetes config.

This is a problem, as people might come with their own kubernetes
cluster and .kube/config file.

This fixes the problem by ensuring that if automation files are
in place, use them (and refresh them if necessary), and fallback
to use local deployer files if necessary.